+++ test/sixarm_ruby_markdown_table_of_contents_test/sixarm/markdown/toc/generator_test.rb
@@ -1,22 +1,34 @@
 # -*- coding: utf-8 -*-
 require "sixarm_ruby_markdown_table_of_contents_test"
 
-describe Markdown::TOC::Generator do
+describe SixArm::Markdown::TOC::Generator do
 
   include Let
 
-  let(:default_settings){ Markdown::TOC::Settings.new }
-  let(:custom_settings){ Markdown::TOC::Settings.new(depth: 3, indent: "!", bullet: "+") }
-  let(:generator_with_default_settings) { Markdown::TOC::Generator.new(settings: default_settings) }
-  let(:generator_with_custom_settings) { Markdown::TOC::Generator.new(settings: custom_settings) }
+  let(:default_settings){ 
+    SixArm::Markdown::TOC::Settings.new 
+  }
+
+  let(:custom_settings){ 
+    SixArm::Markdown::TOC::Settings.new(
+      bullet: custom_bullet,
+      line_prefix: custom_line_prefix, 
+      headline_minimum_level: custom_headline_minimum_level, 
+      headline_maximum_level: custom_headline_maximum_level,
+    )
+  }
+
+  let(:generator_with_default_settings) { SixArm::Markdown::TOC::Generator.new(settings: default_settings) }
+  let(:generator_with_custom_settings) { SixArm::Markdown::TOC::Generator.new(settings: custom_settings) }
+
   let(:s) { markdown_input_as_markdown_string }
-  let(:headline){ Headline.new(level: 3, text: "Hello World") }
+  let(:headline){ SixArm::Markdown::Headline.new(level: 3, text: "Hello World") }
 
   describe "#headline_to_markdown" do
 
     describe "with default settings" do
 
-      it "return the headline, with a default prefix ``, with a two-space indent because the headline is level 3, with a default bullet `*`" do  
+      it "return the headline, with a default line prefix ``, with a two-space indent because the headline is level 3, with a default bullet `*`" do  
         g = generator_with_default_settings
         actual = g.headline_to_markdown(headline)
         expect = "  * [Hello World](#hello-world)\n"
@@ -27,7 +39,7 @@ describe Markdown::TOC::Generator do
 
     describe "with custom settings" do
 
-      it "return the headline, with a custom prefix `!`, with a two-space indent because the headline is level 3, with a custom bullet `+`" do
+      it "return the headline, with a custom line prefix `!`, with a two-space indent because the headline is level 3, with a custom bullet `+`" do
         g = generator_with_custom_settings
         actual = g.headline_to_markdown(headline)
         expect = "!  + [Hello World](#hello-world)\n"
@@ -53,7 +65,14 @@ describe Markdown::TOC::Generator do
 
     describe "with custom settings" do
 
-      it "return the headlines that have a level less than or equal to the settings depth" do
+      it "return the headlines that have a level greater than than or equal to the settings headline minimum level" do
+        g = generator_with_custom_settings
+        actual = g.scan_headlines(s)
+        expect = markdown_custom_output_as_headlines
+        expect(actual).must_equal(expect)
+      end
+
+      it "return the headlines that have a level less than or equal to the settings headline maximum level" do
         g = generator_with_custom_settings
         actual = g.scan_headlines(s)
         expect = markdown_custom_output_as_headlines
@@ -85,7 +104,7 @@ describe Markdown::TOC::Generator do
         expect = markdown_custom_output_toc_as_markdown_string
         expect(actual).must_equal(expect)
       end
-      
+
     end
 
   end